tests/light: extract control flow related testcases into a separate file
