+++ tests/light/functional_tests/filterx/test_filterx.py
@@ -851,158 +851,6 @@ def test_ternary_return_condition_expression_value_without_true_branch(config, s
     assert file_true.read_log() == "boolean:true\n"
 
 
-def test_if_condition_without_else_branch_match(config, syslog_ng):
-    (file_true, file_false) = create_config(
-        config, """
-    $out = "default";
-    if (true) {
-        $out = "matched";
-    };
-    $MSG = $out;
-""",
-    )
-    syslog_ng.start(config)
-
-    assert file_true.get_stats()["processed"] == 1
-    assert "processed" not in file_false.get_stats()
-    assert file_true.read_log() == "matched\n"
-
-
-def test_if_condition_without_else_branch_nomatch(config, syslog_ng):
-    (file_true, file_false) = create_config(
-        config, """
-    $out = "default";
-    if (false) {
-        $out = "matched";
-    };
-    $MSG = $out;
-""",
-    )
-    syslog_ng.start(config)
-
-    assert file_true.get_stats()["processed"] == 1
-    assert "processed" not in file_false.get_stats()
-    assert file_true.read_log() == "default\n"
-
-
-def test_if_condition_no_matching_condition(config, syslog_ng):
-    (file_true, file_false) = create_config(
-        config, """
-    $out = "default";
-    if (false) {
-        $out = "matched";
-    } elif (false) {
-        $out = "elif-matched";
-    };
-    $MSG = $out;
-""",
-    )
-    syslog_ng.start(config)
-
-    assert file_true.get_stats()["processed"] == 1
-    assert "processed" not in file_false.get_stats()
-    assert file_true.read_log() == "default\n"
-
-
-def test_if_condition_matching_main_condition(config, syslog_ng):
-    (file_true, file_false) = create_config(
-        config, """
-    $out = "default";
-    if (true) {
-        $out = "matched";
-    } elif (true) {
-        $out = "elif-matched";
-    } else {
-        $out = "else-matched";
-    };
-    $MSG = $out;
-""",
-    )
-    syslog_ng.start(config)
-
-    assert file_true.get_stats()["processed"] == 1
-    assert "processed" not in file_false.get_stats()
-    assert file_true.read_log() == "matched\n"
-
-
-def test_if_condition_matching_elif_condition(config, syslog_ng):
-    (file_true, file_false) = create_config(
-        config, """
-    $out = "default";
-    if (false) {
-        $out = "matched";
-    } elif (true) {
-        $out = "elif-matched";
-    } elif (true) {
-        $out = "elif2-matched";
-    } else {
-        $out = "else-matched";
-    };
-    $MSG = $out;
-""",
-    )
-    syslog_ng.start(config)
-
-    assert file_true.get_stats()["processed"] == 1
-    assert "processed" not in file_false.get_stats()
-    assert file_true.read_log() == "elif-matched\n"
-
-
-def test_if_condition_matching_else_condition(config, syslog_ng):
-    (file_true, file_false) = create_config(
-        config, """
-    $out = "default";
-    if (false) {
-        $out = "matched";
-    } elif (false) {
-        $out = "elif-matched";
-    } else {
-        $out = "else-matched";
-    };
-    $MSG = $out;
-""",
-    )
-    syslog_ng.start(config)
-
-    assert file_true.get_stats()["processed"] == 1
-    assert "processed" not in file_false.get_stats()
-    assert file_true.read_log() == "else-matched\n"
-
-
-def test_if_condition_matching_expression(config, syslog_ng):
-    (file_true, file_false) = create_config(
-        config, """
-    $out = "default";
-    if ("foo" eq "foo") {
-        $out = "matched";
-    };
-    $MSG = $out;
-""",
-    )
-    syslog_ng.start(config)
-
-    assert file_true.get_stats()["processed"] == 1
-    assert "processed" not in file_false.get_stats()
-    assert file_true.read_log() == "matched\n"
-
-
-def test_if_condition_non_matching_expression(config, syslog_ng):
-    (file_true, file_false) = create_config(
-        config, """
-    $out = "default";
-    if ("foo" eq "bar") {
-        $out = "matched";
-    };
-    $MSG = $out;
-""",
-    )
-    syslog_ng.start(config)
-
-    assert file_true.get_stats()["processed"] == 1
-    assert "processed" not in file_false.get_stats()
-    assert file_true.read_log() == "default\n"
-
-
 def test_isset_existing_value_not_in_scope_yet(config, syslog_ng):
     (file_true, file_false) = create_config(
         config, """
@@ -2268,70 +2116,6 @@ def test_list_range_check_out_of_range(config, syslog_ng):
     syslog_ng.start(config)
     assert file_false.get_stats()["processed"] == 1
     assert file_false.read_log() == "foobar\n"
-
-
-def test_drop(config, syslog_ng):
-    file_true = config.create_file_destination(file_name="dest-true.log", template="'$MSG\n'")
-    file_false = config.create_file_destination(file_name="dest-false.log", template="'$MSG\n'")
-
-    raw_conf = f"""
-@version: {config.get_version()}
-
-options {{ stats(level(1)); }};
-
-source genmsg {{
-    example-msg-generator(num(1) template("foo"));
-    example-msg-generator(num(1) template("bar"));
-}};
-
-destination dest_true {{
-    {render_statement(file_true)};
-}};
-
-destination dest_false {{
-    {render_statement(file_false)};
-}};
-
-log {{
-    source(genmsg);
-    if {{
-        filterx {{ {"if ($MSG =~ 'foo') {drop;};"} \n}};
-        destination(dest_true);
-    }} else {{
-        destination(dest_false);
-    }};
-}};
-"""
-    config.set_raw_config(raw_conf)
-
-    syslog_ng.start(config)
-
-    assert "processed" in file_true.get_stats()
-    assert file_true.get_stats()["processed"] == 1
-    assert file_true.read_log() == 'bar\n'
-    assert syslog_ng.wait_for_message_in_console_log("filterx rule evaluation result; result='explicitly dropped'") != []
-
-
-def test_done(config, syslog_ng):
-    (file_true, file_false) = create_config(
-        config,
-        msg="foo",
-        filterx_expr_1=r"""
-            if ($MSG =~ 'foo') {
-              declare var_wont_change = true;
-              done;
-              var_wont_change = false; # This will be skipped
-            };
-        """,
-        filterx_expr_2=r"""
-            $MSG = vars();
-        """,
-    )
-    syslog_ng.start(config)
-
-    assert file_true.get_stats()["processed"] == 1
-    assert "processed" not in file_false.get_stats()
-    assert file_true.read_log() == '{"$MESSAGE":"foo","var_wont_change":true}\n'
 
 
 def test_parse_xml(config, syslog_ng):
